versão final

+++ discagem_condicao.py
@@ -1,3 +1,5 @@
+from math import floor
+
 menu = 5
 novo = ''
 while menu != 6:
@@ -24,8 +26,8 @@ while menu != 6:
             print(f"A multiplicação dos valores {numero1} X {numero2} = {multiplicacao}")
         case 4:  # divisão
             while numero2 == 0:  #verifica se não é uma divisão por 0
-                numero2 = int(input("Não é possível divisão por 0\n"
-                                    "Informe um novo valor para o 2º Número: "))
+                numero2 = float(input("Não é possível divisão por 0\n"
+                                      "Informe um novo valor para o 2º Número: "))
             divisao = numero1 / numero2
             print(f"A divisão dos valores {numero1} / {numero2} = {divisao}")
         case 5:  #repete para inserir um novo número
